requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
anthropic==0.26.0
pydantic==2.7.1
pydantic-settings==2.2.1
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.0
openai==1.30.0
playwright==1.44.0
bcrypt==4.1.3
PyJWT==2.8.0
razorpay==1.4.1
google-auth==2.29.0
psutil==5.9.8
reportlab==4.1.0
asyncpg==0.29.0
redis==5.0.4
hiredis==2.3.2
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
stripe==9.12.0
pytz==2024.1
